+++ lending/src/sdk/models/shared/accountingcustomers.ts
@@ -7,9 +7,6 @@ import { AccountingCustomer } from "./accountingcustomer";
 import { Links } from "./links";
 import { Expose, Type } from "class-transformer";
 
-/**
- * Success
- */
 export class AccountingCustomers extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "_links" })