+++ lending/src/sdk/models/shared/accountingjournalentries.ts
@@ -7,9 +7,6 @@ import { AccountingJournalEntry } from "./accountingjournalentry";
 import { Links } from "./links";
 import { Expose, Type } from "class-transformer";
 
-/**
- * Success
- */
 export class AccountingJournalEntries extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "_links" })