+++ lending/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v0.1.0] lending
 ### Releases
 - [NPM v0.1.0] https://www.npmjs.com/package/@codat/lending/v/0.1.0 - lending
+
+## 2023-09-13 12:15:13
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.83.0 (2.109.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.1.0] lending
+### Releases
+- [NPM v1.1.0] https://www.npmjs.com/package/@codat/lending/v/1.1.0 - lending